Support floating-point scores for tie-split points (e.g. 20.5/20.5)

+++ yraa/io.py
@@ -33,7 +33,7 @@ def load_results_from_csv(path: str) -> List[RaceResult]:
                 cell = cell.strip()
                 if cell == "":
                     continue
-                score = int(cell)
+                score = float(cell)
                 results.append(
                     RaceResult(
                         athlete_name=athlete_name,
